Some correctings for using the GlobalAttachments functions. Now synapse detection and function working.

+++ ugbase/lib_grid/algorithms/geom_obj_util/vertex_util_impl.hpp
@@ -387,7 +387,7 @@ int FindVertexByCoordinate(const MathVector<dim>& coord, size_t ncoords, const M
 
 	if (ncoords <= 0) return -1;
 
-	int bestVrt = 0;
+	size_t bestVrt = 0;
 	number bestDistSq = VecDistanceSq(coord, vCoords[0]);
 
 	for (int i=1; i<ncoords; ++i)
